Add support for grunt
