+++ .gitignore
@@ -70,5 +70,9 @@ fabric.properties
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
 
+# Node
+node_modules/
+package-lock.json
+
 # Ignore extracted font assets files, but keep the zip
 src/assets/tmp/glyphter-font/
