+++ IEconfig.xml
@@ -2,10 +2,10 @@
 <browserconfig>
     <msapplication>
         <tile>
-            <square70x70logo src="./src/assets/icons/mstile-70x70.png"/>
-            <square150x150logo src="./src/assets/icons/mstile-150x150.png"/>
-            <square310x310logo src="./src/assets/icons/mstile-310x310.png"/>
-            <wide310x150logo src="./src/assets/icons/mstile-310x150.png"/>
+            <square70x70logo src="dist/assets/mstile-70x70.png"/>
+            <square150x150logo src="dist/assets/mstile-150x150.png"/>
+            <square310x310logo src="dist/assets/mstile-310x310.png"/>
+            <wide310x150logo src="dist/assets/mstile-310x150.png"/>
             <TileColor>#58D5F6</TileColor>
         </tile>
     </msapplication>
