+++ README.md
@@ -10,6 +10,19 @@ There have been many improvements:
 - Additional filters
 - Full suite of GBGW font glyphs
 
+## Installation
+
+This project uses [npm](https://www.npmjs.com/) and [grunt](https://gruntjs.com/) to deploy bundle files to improve load and performance.
+
+You will need to install [Node.js](https://nodejs.org/en/download/) to deploy distribution files.
+
+You can install and build using the following commands:
+
+```shell script
+npm install
+npm run build
+```
+
 ## Changelog
 
 For version history, please view the [change log](./CHANGELOG.md).
