+++ index.html
@@ -8,31 +8,25 @@
     <meta name="keywords" content="GBGW, Gundam, Gundam Battle Gunpla Warfare, Calculator">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="language" content="ES6">
-    <meta name="msapplication-config" content="./IEconfig.xml">
-    <link rel="icon" type="image/png" sizes="16x16" href="./src/assets/icons/favicon-16x16.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="./src/assets/icons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="96x96" href="./src/assets/icons/favicon-96x96.png">
-    <link rel="apple-touch-icon" href="./src/assets/icons/apple-iphone-old.png">
-    <link rel="apple-touch-icon" sizes="180x180" href="./src/assets/icons/apple-touch-icon.png">
-    <link rel="apple-touch-icon" sizes="152x152" href="./src/assets/icons/apple-ipad-retina.png">
-    <link rel="apple-touch-icon" sizes="167x167" href="./src/assets/icons/apple-ipad-pro.png">
-    <link rel="manifest" href="./site.webmanifest">
+    <meta name="msapplication-config" content="IEconfig.xml">
+    <link rel="icon" type="image/png" sizes="16x16" href="dist/assets/favicon-16x16.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="dist/assets/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="dist/assets/favicon-96x96.png">
+    <link rel="apple-touch-icon" href="dist/assets/apple-iphone-old.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="dist/assets/apple-touch-icon.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="dist/assets/apple-ipad-retina.png">
+    <link rel="apple-touch-icon" sizes="167x167" href="dist/assets/apple-ipad-pro.png">
+    <link rel="manifest" href="site.webmanifest">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.min.css"/>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.4/balloon.min.css"/>
-    <link rel="stylesheet" href="./src/css/gbgw-icons.css"/>
-    <link rel="stylesheet" href="./src/css/style.css"/>
-    <link rel="stylesheet" href="./src/css/themes/dark.css"/>
-    <script src="./src/js/globals.js" type="text/javascript"></script>
-    <script src="./src/js/classes.js" type="text/javascript"></script>
-    <script src="./src/js/calculator.js" type="text/javascript"></script>
-    <script src="./src/js/data.js" type="text/javascript"></script>
-    <script src="./src/js/main.js" type="text/javascript"></script>
+    <link rel="stylesheet" href="dist/css/bundle.min.css"/>
+    <script src="dist/js/bundle.min.js" type="text/javascript"></script>
   </head>
   <body>
     <div class="app">
       <header>
         <div class="options">
-          <h1>GBGW Calculator v1.2.0</h1>
+          <h1>GBGW Calculator v1.2.1</h1>
           <div class="filters">
             <span>Filters</span>
             <select class="unselected-filter js-filter-attr">
@@ -90,27 +84,27 @@
         <div class="brands">
           <div class="brand-badge-wrapper" title="BANDAI NAMCO Entertainment">
             <a href="https://www.bandainamcoent.com" target="_blank">
-              <img src="./src/assets/third-party/bandai-namco-logo.svg" height="24em" alt="BANDAI NAMCO Entertainment"/>
+              <img src="dist/assets/bandai-namco-logo.svg" height="24em" alt="BANDAI NAMCO Entertainment"/>
             </a>
           </div>
           <div class="brand-badge-wrapper" title="Bandai Co., Ltd.">
             <a href="https://www.bandai.com/" target="_blank">
-              <img src="./src/assets/third-party/bandai-logo.svg" height="24em" alt="Bandai Co., Ltd."/>
+              <img src="dist/assets/bandai-logo.svg" height="24em" alt="Bandai Co., Ltd."/>
             </a>
           </div>
           <div class="brand-badge-wrapper" title="Gundam Battle: Gunpla Warfare">
             <a href="https://g-b-en.ggame.jp" target="_blank">
-              <img src="./src/assets/third-party/gbm-logo.jpg" height="24em" alt="Gundam Battle: Gunpla Warfare"/>
+              <img src="dist/assets/gbm-logo.jpg" height="24em" alt="Gundam Battle: Gunpla Warfare"/>
             </a>
           </div>
           <div class="brand-badge-wrapper">
             <a href="https://discord.gg/6aMUnBD" title="GBGW: Newtype Labs – Discord" target="_blank">
-              <img src="./src/assets/third-party/newtype-labs-logo.svg" height="24em" alt="Gundam Newtype Labs"/>
+              <img src="dist/assets/newtype-labs-logo.svg" height="24em" alt="Gundam Newtype Labs"/>
             </a>
           </div>
           <div class="brand-badge-wrapper">
             <a href="https://github.com/about" title="GitHub, Inc." target="_blank">
-              <img src="./src/assets/third-party/github-logo.svg" height="24em" alt="GitHub, Inc."/>
+              <img src="dist/assets/github-logo.svg" height="24em" alt="GitHub, Inc."/>
             </a>
           </div>
         </div>
