+++ site.webmanifest
@@ -2,11 +2,11 @@
     "name": "GBGW Calculator",
     "short_name": "GBGW Calc",
     "icons": [{
-        "src": "./src/assets/icons/android-chrome-192x192.png",
+        "src": "dist/assets/android-chrome-192x192.png",
         "sizes": "192x192",
         "type": "image/png"
     }, {
-        "src": "./src/assets/icons/android-chrome-512x512.png",
+        "src": "dist/assets/android-chrome-512x512.png",
         "sizes": "512x512",
         "type": "image/png"
     }],
